airport_driver/requirements.txt
certifi==2025.7.14
charset-normalizer==3.4.2
filelock==3.18.0
idna==3.10
pycomm3==1.2.14
pygame==2.6.1
pymodbus==3.10.0
python-snap7==2.0.2
requests==2.32.4
screeninfo==0.8.1
urllib3==2.5.0
